+++ packages/firestore/CHANGELOG.md
@@ -1,4 +1,11 @@
-# Unreleased
+# Unreleased (firestore-api-changes)
+- [changed] Snapshot listeners (with the `includeMetadataChanges` option
+  enabled) now receive an event with `snapshot.metadata.fromCache` set to
+  `true` if the SDK loses its connection to the backend. A new event with
+  `snapshot.metadata.fromCache` set to false will be raised once the
+  connection is restored and the query is in sync with the backend again.
+
+# v0.2.0
 - [feature] Added Node.js support for Cloud Firestore (with the exception of
   the offline persistence feature).
 - [changed] Webchannel requests use $httpHeaders URL parameter rather than